requirements-backend.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
google-genai==1.10.0
python-dotenv==1.2.1
PyPDF2==3.0.1

# Auth dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
authlib==1.3.2
httpx>=0.28.1
itsdangerous==2.2.0
email-validator>=2.0.0

# Database
sqlalchemy==2.0.36
